Submits code to server and receives result
Bring in JQuery, its utterly obnoxious to do ajax requests w/o it.
Remove the EvalIn wrapper, I wasn't sure what its purpose was
(it does catch an error that maybe was getting raised?)
IDK, we can add it back in if it turns out there's a use-case I'm just
ignorant of.

This progresses #12

Closes #14
Closes #15 <-- this was actually done in 410ce57dade43e6be2e3bd1249c374fb85e512cd